Moved form components to general component library and cleaned up usage
