+++ assets/src/edit-story/components/form/index.js
@@ -14,14 +14,5 @@
  * limitations under the License.
  */
 
-/**
- * External dependencies
- */
-import styled from 'styled-components';
-
-const ActionButton = styled.button.attrs( { type: 'button' } )`
-	color: ${ ( { theme } ) => theme.colors.mg.v1 };
-	font-size: 11px;
-`;
-
-export default ActionButton;
+export { default as InputGroup } from './inputGroup';
+export { default as SelectMenu } from './select';
