+++ assets/src/edit-story/components/form/label.js
