+++ assets/src/edit-story/panels/backgroundColor.js
@@ -28,8 +28,8 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
+import { InputGroup } from '../components/form';
 import { SimplePanel } from './panel';
-import { InputGroup } from './components';
 import getCommonValue from './utils/getCommonValue';
 
 function BackgroundColorPanel( { selectedElements, onSetProperties } ) {
