+++ assets/src/edit-story/panels/panel/title.js
@@ -96,7 +96,7 @@ function Title( { children, isPrimary, isResizable } ) {
 		[ setHeight, maxHeight ],
 	);
 
-	const ariaLabel = isCollapsed ? __( 'Expand panel', 'web-stories' ) : __( 'Collapse panel', 'web-stories' );
+	const titleLabel = isCollapsed ? __( 'Expand panel', 'web-stories' ) : __( 'Collapse panel', 'web-stories' );
 
 	return (
 		<Header isPrimary={ isPrimary } hasResizeHandle={ isResizable && ! isCollapsed }>
@@ -110,7 +110,7 @@ function Title( { children, isPrimary, isResizable } ) {
 			) }
 			<HeaderButton
 				onClick={ isCollapsed ? expand : collapse }
-				aria-label={ ariaLabel }
+				aria-label={ titleLabel }
 				aria-expanded={ ! isCollapsed }
 				aria-controls={ panelContentId }
 			>
