+++ assets/src/edit-story/panels/size.js
@@ -28,8 +28,8 @@ import { __, _x } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
+import { InputGroup } from '../components/form';
 import { SimplePanel } from './panel';
-import { InputGroup } from './components';
 import getCommonValue from './utils/getCommonValue';
 
 function SizePanel( { selectedElements, onSetProperties } ) {
