+++ assets/src/edit-story/panels/style.js
@@ -28,8 +28,8 @@ import { __, _x } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
+import { InputGroup, SelectMenu } from '../components/form';
 import { SimplePanel } from './panel';
-import { SelectMenu, InputGroup } from './components';
 import getCommonValue from './utils/getCommonValue';
 
 function StylePanel( { selectedElements, onSetProperties } ) {
